+++ app/Core/routes/components/filemanager.route.php
