+++ app/Core/routes/console.php
@@ -1,4 +1,13 @@
 <?php
+/**
+ * MYMO CMS - API Routes
+ *
+ * @package mymocms/mymocms
+ * @author The Anh Dang
+ *
+ * Developed based on Laravel Framework
+ * Github: https://github.com/mymocms/mymocms
+ */
 
 use Illuminate\Foundation\Inspiring;
