+++ artisan
@@ -1,5 +1,14 @@
 #!/usr/bin/env php
 <?php
+/**
+ * MYMO CMS
+ *
+ * @package mymocms/mymocms
+ * @author The Anh Dang
+ *
+ * Developed based on Laravel Framework
+ * Github: https://github.com/mymocms/mymocms
+*/
 
 define('LARAVEL_START', microtime(true));
