+++ modules/Movie/Models/BlockIp/CountryName.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace App\Core\Models\BlockIp;
+namespace Modules\Movie\Models\BlockIp;
 
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * App\Core\Models\BlockIp\CountryName
+ * Modules\Movie\Models\BlockIp\CountryName
  *
  * @property int $id
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName query()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\BlockIp\CountryName whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\BlockIp\CountryName whereUpdatedAt($value)
  * @mixin \Eloquent
  */
 class CountryName extends Model
