+++ modules/Movie/Models/Category/Countries.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App\Core\Models\Category;
+namespace Modules\Movie\Models\Category;
 
 use App\Core\Traits\UseMetaSeo;
 use App\Core\Traits\UseSlug;
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * App\Core\Models\Category\Countries
+ * Modules\Movie\Models\Category\Countries
  *
  * @property int $id
  * @property string $name
@@ -15,24 +15,24 @@ use Illuminate\Database\Eloquent\Model;
  * @property int $status
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries query()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereName($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereSlug($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereStatus($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereSlug($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereStatus($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereUpdatedAt($value)
  * @mixin \Eloquent
  * @property string|null $description
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereDescription($value)
  * @property string|null $meta_title
  * @property string|null $meta_description
  * @property string|null $keywords
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereKeywords($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereMetaDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Countries whereMetaTitle($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereKeywords($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereMetaDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Countries whereMetaTitle($value)
  */
 class Countries extends Model
 {
