+++ modules/Movie/Models/Category/Genres.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Core\Models\Category;
+namespace Modules\Movie\Models\Category;
 
 use App\Core\Traits\UseMetaSeo;
 use App\Core\Traits\UseSlug;
@@ -8,7 +8,7 @@ use App\Core\Traits\UseThumbnail;
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * App\Core\Models\Category\Genres
+ * Modules\Movie\Models\Category\Genres
  *
  * @property int $id
  * @property string|null $thumbnail
@@ -18,24 +18,24 @@ use Illuminate\Database\Eloquent\Model;
  * @property int $status
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres query()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereCreatedAt($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereName($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereSlug($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereStatus($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereThumbnail($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereSlug($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereStatus($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereThumbnail($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereUpdatedAt($value)
  * @mixin \Eloquent
  * @property string|null $meta_title
  * @property string|null $meta_description
  * @property string|null $keywords
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereKeywords($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereMetaDescription($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Core\Models\Category\Genres whereMetaTitle($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereKeywords($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereMetaDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\Modules\Movie\Models\Category\Genres whereMetaTitle($value)
  */
 class Genres extends Model
 {
