+++ modules/Movie/Models/PaidMembers/UserSubscription.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Core\Models\PaidMembers;
+namespace Modules\Movie\Models\PaidMembers;
 
 use Illuminate\Database\Eloquent\Model;
 
 /**
- * App\Core\Models\PaidMembers\UserSubscription
+ * Modules\Movie\Models\PaidMembers\UserSubscription
  *
  * @property int $id
  * @property string $token
